collector instrum
