+++ docker-compose.yml
@@ -3,7 +3,6 @@ version: '3'
 services:
   greetings-backend:
     container_name: greetings-backend
-    restart: 'on-failure'
     build:
       context: .
       dockerfile: ./Dockerfile
@@ -11,7 +10,7 @@ services:
       - 80:80
     environment:
       - NAME="VictorHJ"
-      - OTEL_EXPORTER_OTLP_ENDPOINT=http://jaeger:4317
+      - OTEL_LOG_LEVEL=ERROR
 
   prometheus:
     image: prom/prometheus:latest
@@ -26,16 +25,8 @@ services:
   jaeger:
     image: jaegertracing/all-in-one:1.49
     ports:
-      - "6831:6831/udp"
-      - "6832:6832/udp"
-      - "5778:5778"
       - "16686:16686"
-      - "4317:4317"
-      - "4318:4318"
       - "14250:14250"
-      - "14268:14268"
-      - "14269:14269"
-      - "9411:9411"
     environment:
       - LOG_LEVEL=debug
       - COLLECTOR_ZIPKIN_HOST_PORT=9411
