+++ requirements.txt
@@ -2,5 +2,5 @@ fastapi>=0.61,<0.70.0
 uvicorn>=0.11.5,<0.12.0
 opentelemetry-distro==0.41b0
 opentelemetry-exporter-otlp-proto-grpc
+opentelemetry-exporter-otlp-proto-http
 opentelemetry-instrumentation-fastapi
-prometheus-client
